refactor: extract MCP and HTTP plumbing into mcpserver.go
