+++ cmd/critical-thinking/mcpserver_test.go
